Add files via upload
There is a garage where the access road can accommodate any number of trucks
at one time. The garage is build such a way that only the last truck entered can be
moved out. Each of the trucks is identified by a positive integer (a truck_id). Write a
program to handle truck moves, allowing for the following commands:
a) On_road (truck_id);
b) Enter_garage (truck_ id);
c) Exit_garage (truck_id);
d) Show_trucks (garage or road);
If an attempt is made to get out a truck which is not the closest to the garage entry, the error
message Truck x not near garage door.